# lab8_UDP_RAFT/service_2/routes.py
# Import necessary modules
from flask import request, jsonify
from models.database import db
from models.electro_scooter import ElectroScooter
from __main__ import app, server
from app import *

@app.route(f'/credentials', methods=['GET'])
def credentials():
    if request.method == 'GET':
        if server.role == "leader":
            print(server.followers)
            return jsonify(server.followers), 200
        else:
            print(server.leader_data)
            return jsonify(server.leader_data), 200

@app.route('/api/electro-scooters', methods=['POST'])
def create_electro_scooter():
    if server.role != "leader" and ("Token" not in request.headers or request.headers["Token"] != "token"):
        return jsonify({"error": "Only the leader can create a new Electro Scooter"}), 401
    else:
        try:
            # Get data from the request body
            data = request.get_json()
            # Validate and extract required parameters
            name = data['name']
            battery_level = data['battery_level']
            # Create a new Electro Scooter
            electro_scooter = ElectroScooter(name=name, battery_level=battery_level)
            print(electro_scooter.name)
            # Add the Electro Scooter to the database
            db.session.add(electro_scooter)
            db.session.commit()
            
            if server.role == "leader":
                print(server.followers)
                for follower in server.followers:
                    print(follower)
                    requests.post(f"http://{follower['host']}:{follower['port']}" + "/api/electro-scooters", json=data, headers={"Token": "token"})
                    
            return jsonify({"message": "Electro Scooter created successfully"}), 201

        except KeyError:
            return jsonify({"error": "Invalid request data"}), 400
    
@app.route('/api/electro-scooters/<int:scooter_id>', methods=['GET'])
def get_electro_scooter_by_id(scooter_id):
    # Find the Electro Scooter by ID
    scooter = ElectroScooter.query.get(scooter_id)
    if scooter is not None:
        return jsonify({
        "id": scooter.id,
        "name": scooter.name,
        "battery_level": scooter.battery_level
        }), 200
    else:
        return jsonify({"error": "Electro Scooter not found"}), 404
    
@app.route('/api/electro-scooters/<int:scooter_id>', methods=['PUT'])
def update_electro_scooter(scooter_id):
    if server.role != "leader" and ("Token" not in request.headers or request.headers["Token"] != "token"):
        return jsonify({"error": "Only the leader can create a new Electro Scooter"}), 401
    else:
        try:
            # Find the Electro Scooter by ID
            scooter = ElectroScooter.query.get(scooter_id)
            if scooter is not None:
                # Get data from the request body
                data = request.get_json()
                # Update the Electro Scooter properties
                scooter.name = data.get('name', scooter.name)
                scooter.battery_level = data.get('battery_level', scooter.battery_level)
                db.session.commit()

                if server.role == "leader":
                    print(server.followers)
                    for follower in server.followers:
                        print(follower)
                        requests.put(f"http://{follower['host']}:{follower['port']}" + "/api/electro-scooters/"+str(scooter_id), json=data, headers={"Token": "token"})

                return jsonify({"message": "Electro Scooter updated successfully"}), 200
            else:
                return jsonify({"error": "Electro Scooter not found"}), 404
        except Exception as e:
            return jsonify({"error": str(e)}), 500
    
@app.route('/api/electro-scooters/<int:scooter_id>', methods=['DELETE'])
def delete_electro_scooter(scooter_id):
    if server.role != "leader" and ("Token" not in request.headers or request.headers["Token"] != "token"):
        return jsonify({"error": "Only the leader can create a new Electro Scooter"}), 401
    else:
        try:
            # Find the Electro Scooter by ID
            scooter = ElectroScooter.query.get(scooter_id)
            if scooter is not None:
                # Get the password from the request headers
                password = request.headers.get('X-Delete-Password')
            # Check if the provided password is correct
                if password == 'your_secret_password': # Replace with your actual password
                    db.session.delete(scooter)
                    db.session.commit()

                    if server.role == "leader":
                        print(server.followers)
                        for follower in server.followers:
                            print(follower)
                            requests.delete(f"http://{follower['host']}:{follower['port']}" + "/api/electro-scooters/"+ str(scooter_id), headers={"Token": "token", "X-Delete-Password": "your_secret_password"})
                    return jsonify({"message": "Electro Scooter deleted successfully"}), 200
                else:
                    return jsonify({"error": "Incorrect password"}), 401
                
            else:
                return jsonify({"error": "Electro Scooter not found"}), 404
        except Exception as e:
            return jsonify({"error": str(e)}), 500